refactor: created a common function and used it in the files where it is required
